+++ govc/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+### unreleased
+
+* Add `-rescan` option to host.storage.info command
+
 ### 0.6.0 (2016-04-29)
 
 * Add folder commands: info, create, destroy, rename, moveinto